refactor: OotdService.getOotds 로직 OotdDtoAssembler로 분리
